+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 ## [Unreleased]
 
 - bumped cecilifier version to 1.60.0
+- fixed NRE when handling explicit delegate instantiation (#190)
+- fixed TypeAttributes for static top level/inner types (#191)
+- change code to copy .runtimeconfig.json file when running the cecilified code (#189)
+- do not load implicit 'this' for static property access (#187)
 
 ## 15/Sept/2022